xserver          commit 080fb49eff4de7ec3a29214994d1403e4d877f6a libxcb           commit daf2c5397607c32c3814f02831b9c1158fd1e52c xkeyboard-config commit 5f12a603e2936a4aa3ea4c08f000392a0d01adae libX11           commit ba15d51bfe789e84e291ccfb31d5f0ae6192969e libXdmcp         commit a3abc2d634c315e7776e5500b2d68f4eed6c2e20 libXau           commit a9c65683e68b3a4349afee5d7673b393fb924d2e xkbcomp          commit a9c3d35a1f74de1681aeb7cdd66dce7bf186c666 pixman           commit 707d7e34ca3d435b23ac3e7c05417fc0d39f4263 fontconfig       commit 127bb4dbfca7ba669eb40212ffcf24dd8f30351b mesa             commit 70767b9cd47df498ed510ec63becd67649518b7f putty            commit 5b2fa80c5758fd52b034dd5daac45604285447b3 xorgproto        commit e312ab76c4d93f2e637ac0c73987b22cd43420c7

+++ libXdmcp/Write.c
@@ -40,7 +40,7 @@ XdmcpWriteHeader (
 {
     BYTE    *newData;
 
-    if ((int)buffer->size < 6 + (int)header->length)
+    if (buffer->size < 6 + (int)header->length)
     {
 	newData = calloc(XDM_MAX_MSGLEN, sizeof (BYTE));
 	if (!newData)
